Remove the ui extension factory.
